add qsv support

+++ ffmpegcv/video_info.py
@@ -6,6 +6,12 @@ import xml.etree.ElementTree as ET
 
 scan_the_whole = {'mkv', 'flv', 'ts'} #scan the whole file to the count, slow
 
+_inited_get_num_NVIDIA_GPUs = False
+_inited_get_num_QSV_GPUs = False
+_num_NVIDIA_GPUs = -1
+_num_QSV_GPUs = -1
+
+
 def get_info(video:str):
     do_scan_the_whole = video.split('.')[-1] in scan_the_whole
     use_count_packets = '-count_packets' if do_scan_the_whole else ''
@@ -33,16 +39,30 @@ def get_info(video:str):
 
 
 def get_num_NVIDIA_GPUs():
-    cmd = "ffmpeg -f lavfi -i nullsrc -c:v h264_nvenc -gpu list -f null -"
-    p = Popen(cmd.split(), shell=False, stdin=PIPE, stdout=PIPE, stderr=PIPE)
-    stdout, stderr = p.communicate(b"")
-    p.stdin.close()
-    p.stdout.close()
-    p.terminate()
-    pattern = re.compile(r"GPU #\d+ - < ")
-    nv_info = pattern.findall(stderr.decode())
-    num_gpu = len(nv_info)
-    return num_gpu
+    global _num_NVIDIA_GPUs, _inited_get_num_NVIDIA_GPUs
+    if not _inited_get_num_NVIDIA_GPUs:
+        cmd = "ffmpeg -f lavfi -i nullsrc -c:v h264_nvenc -gpu list -f null -"
+        p = Popen(cmd.split(), shell=False, stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        stdout, stderr = p.communicate(b"")
+        p.stdin.close()
+        p.stdout.close()
+        p.terminate()
+        pattern = re.compile(r"GPU #\d+ - < ")
+        nv_info = pattern.findall(stderr.decode())
+        _num_NVIDIA_GPUs = len(nv_info)
+        _inited_get_num_NVIDIA_GPUs = True
+    return _num_NVIDIA_GPUs
+
+
+def get_num_QSV_GPUs():
+    global _num_QSV_GPUs, _inited_get_num_QSV_GPUs
+    if not _inited_get_num_QSV_GPUs:
+        cmd = "ffmpeg -hide_banner -f qsv -h encoder=h264_qsv"
+        p = Popen(cmd.split(), shell=False, stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        stdout, stderr = p.communicate(b"")
+        _num_QSV_GPUs = 1 if len(stdout) > 50 else 0
+        _inited_get_num_QSV_GPUs = True
+    return _num_QSV_GPUs
 
 
 def encoder_to_nvidia(codec):
@@ -54,6 +74,22 @@ def encoder_to_nvidia(codec):
         return codec
     else:
         raise Exception("No NV codec found for %s" % codec)
+
+
+def encoder_to_qsv(code):
+    codec_map = {
+        "h264": "h264_qsv",
+        "hevc": "hevc_qsv",
+        "mjpeg": "mjpeg_qsv",
+        "mpeg2video": "mpeg2_qsv",
+        "vp9": "vp9_qsv"}
+
+    if codec in codec_map:
+        return codec_map[codec]
+    elif codec in codec_map.values():
+        return codec
+    else:
+        raise Exception("No QSV codec found for %s" % codec)
 
 
 def decoder_to_nvidia(codec):
@@ -79,6 +115,25 @@ def decoder_to_nvidia(codec):
         return codec
     else:
         raise Exception("No NV codec found for %s" % codec)
+
+
+def decoder_to_qsv(code):
+    codec_map = {
+        "av1": "av1_qsv",
+        "h264": "h264_qsv",
+        "hevc": "hevc_qsv",
+        "mjpeg": "mjpeg_qsv",
+        "mpeg2video": "mpeg2_qsv",
+        "vc1": "vc1_qsv",
+        "vp8": "vp8_qsv",
+        "vp9": "vp9_qsv"}
+
+    if codec in codec_map:
+        return codec_map[codec]
+    elif codec in codec_map.values():
+        return codec
+    else:
+        raise Exception("No QSV codec found for %s" % codec)
 
 
 def run_async(args):
